state-machine: fix broken logic (see below)
it's interesting how much I am learning

the logic here on the surface looked completely correct and elegant, but it was wrong

what happens if the miner is also the from account? or the miner is the to account? or the from and to accounts are the same?

in the current code, it would basically return 3 state leaves (from, to, miner) - where 2 of those leaves would be the same account, but with different balances

the new code is a v1. It might do the job, for now it's just MVP functionality while I'm writing the TestStateMachineTxAlreadySequenced test.

I will write some proper tests in the state machine + fuzzing later.